+++ src/lib/datatransfers/scripts/job_s3_downloader.sh
@@ -26,7 +26,13 @@ do
             part_i=1
             headers_file=$(mktemp)
 
-            if [[ -e "$target_file" ]]; then rm "$target_file"; fi
+            if [ -d "$target_file" ]; then
+
+                echo $(date -u) "targetPath must point to a file path: $target_file is a directory." >&2
+                exit 3
+            fi
+
+            if [[ -f "$target_file" ]]; then rm "$target_file"; fi
 
             while [[ "$download_complete" != "true" ]];
             do