+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.4.2] - OPEN
 
 ### Added
+- Fine grained probing services per cluster.
 
 ### Changed