+++ CHANGELOG.md
@@ -8,13 +8,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.4.2] - OPEN
 
 ### Added
-- Fine grained probing services per cluster.
+- File target check in S3 file transfer job for external file upload.
+
+- `cluster.scheduler.connection_mode` setting to configure how the client connects to the scheduler backend (`ssh`, `rest` or `hybrid`)
+- Get jobs now allows to specify the account parameter
 
 ### Changed
 
-### Fixed
+- Large file download via s3 no loger appends uuid to file name.
 
-## [2.4.1] 
+### Fixed
+- File transfer examples with .NET
+
+## [2.4.1]
 
 ### Added
 
@@ -36,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Timeout value for commands executed via `filesystems/ops` and `status` is now configured with command execution timeout setting
 - Stdout, stderr path are now fully expanded
 - `probing` configuration is optional now for the `clusters` settings
+- Updated documentation for large data upload
 
 ## [2.4.0]